Fix strict cache methods
