+++ Cartfile.resolved
@@ -1,3 +1,3 @@
-github "krzyzanowskim/CryptoSwift" "0.6.1"
-github "Quick/Nimble" "v5.1.0"
-github "Quick/Quick" "v0.10.0"
+github "krzyzanowskim/CryptoSwift" "0.6.8"
+github "Quick/Nimble" "v6.1.0"
+github "Quick/Quick" "v1.1.0"
