+++ Source/Shared/HybridCache.swift
@@ -1,13 +1,9 @@
-import Foundation
-
 /**
- Cache is type safe cache, works with objects that conform to
- Cachable protocol. It's two layered cache (with front and back storages).
+ HybridCache supports storing all kinds of objects, as long as they conform to
+ Cachable protocol. It's two layered cache (with front and back storages), as well as Cache.
  Subscribes to system notifications to clear expired cached objects.
  */
-public final class Cache<T: Cachable>: BasicHybridCache {
-
-  // MARK: - Caching
+public class HybridCache: BasicHybridCache {
 
   /**
    Adds passed object to the front and back cache storages.
@@ -17,7 +13,7 @@ public final class Cache<T: Cachable>: BasicHybridCache {
    - Parameter expiry: Expiration date for the cached object
    - Parameter completion: Completion closure to be called when the task is done
    */
-  public func add(_ key: String, object: T, expiry: Expiry? = nil, completion: (() -> Void)? = nil) {
+  public override func add<T: Cachable>(_ key: String, object: T, expiry: Expiry? = nil, completion: (() -> Void)? = nil) {
     super.add(key, object: object, expiry: expiry, completion: completion)
   }
 
@@ -27,8 +23,7 @@ public final class Cache<T: Cachable>: BasicHybridCache {
    - Parameter key: Unique key to identify the object in the cache
    - Parameter completion: Completion closure returns object or nil
    */
-
-  public func object(_ key: String, completion: @escaping (_ object: T?) -> Void) {
+  public override func object<T: Cachable>(_ key: String, completion: @escaping (_ object: T?) -> Void) {
     super.object(key, completion: completion)
   }
 }
